Dev branch for 0.9.0

Closes #1598, #1969, #1982, #1974, #1975, #2000, and #1140

See merge request plom/plom!1322

+++ plom/create/buildClasslist.py
@@ -1,41 +1,26 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
-# Copyright (C) 2018-2020 Andrew Rechnitzer
+# Copyright (C) 2018-2022 Andrew Rechnitzer
 # Copyright (C) 2019-2022 Colin B. Macdonald
 # Copyright (C) 2020 Vala Vakilian
 # Copyright (C) 2020 Dryden Wiebe
 
 import csv
+import importlib.resources as resources
 from pathlib import Path
-import sys
 import tempfile
-
-if sys.version_info >= (3, 7):
-    import importlib.resources as resources
-else:
-    import importlib_resources as resources
 
 import pandas
 
 import plom
 from plom.finish.return_tools import import_canvas_csv
 
-
-possible_surname_fields = ["surname", "familyName", "lastName"]
-
-possible_given_name_fields = [
-    "name",
-    "givenName",
-    "firstName",
-    "givenNames",
-    "firstNames",
-    "preferredName",
-    "preferredNames",
-    "nickName",
-    "nickNames",
-]
-
-canvas_columns_format = ("Student", "ID", "SIS User ID", "SIS Login ID")
-
+from plom.create.classlistValidator import (
+    possible_sid_fields,
+    possible_one_name_fields,
+    possible_given_name_fields,
+    possible_surname_fields,
+    PlomClasslistValidator,
+)
 
 # Note: file is full of pandas warnings, which I think are false positives
 # pylint: disable=unsubscriptable-object
@@ -61,19 +46,41 @@ def clean_non_canvas_csv(csv_file_name):
     print('Extracting columns from csv file: "{0}"'.format(csv_file_name))
 
     # strip excess whitespace from column names
-    df.rename(columns=lambda x: x.strip(), inplace=True)
+    df.rename(columns=lambda x: str(x).strip(), inplace=True)
 
-    if "id" not in df.columns:
+    # find the id column
+    id_column = None
+    for c in df.columns:
+        if c.casefold() in (x.casefold() for x in possible_sid_fields):
+            # print(f'"{c}" column present')
+            id_column = c
+            break
+    if id_column is None:
+        # note that this should be caught by the validator
         raise ValueError('no "id" column is present')
-    print('"id" column present')
-    # strip excess whitespace
-    df["id"] = df["id"].apply(lambda X: X.strip())
+    # make sure id column named 'id' - lowercase
+    print(f"Renaming column {id_column} to 'id'")
+    df.rename(columns={id_column: "id"}, inplace=True)
+    # clean up the column - strip whitespace
+    df["id"] = df["id"].apply(lambda X: str(X).strip())  # avoid issues with non-string
+    # print('"id" column present')
 
-    # see if we have our preferred format for a name column
-    if "studentName" in df.columns:
-        print('"studentName" column present')
-        df["studentName"].apply(lambda X: X.strip())
-        return df
+    # see if there is a single name column
+    one_name_column = None
+    for c in df.columns:
+        if c.casefold() in (x.casefold() for x in possible_one_name_fields):
+            # print(f'"{c}" column present')
+            one_name_column = c
+            break
+    if one_name_column is not None:
+        # make sure id column named 'id' - lowercase
+        print(f"Renaming column {one_name_column} to 'studentName'")
+        df.rename(columns={one_name_column: "studentName"}, inplace=True)
+        # clean up the column - strip whitespace
+        df["studentName"].apply(lambda X: str(X).strip())  # avoid errors with blanks
+        # print('"studentName" column present')
+        # now return only the id and studentName columns
+        return df[["id", "studentName"]]
 
     # Otherwise, we will check the column headers again taking the first
     # columns that looks like firstname and a lastname fields
@@ -81,12 +88,12 @@ def clean_non_canvas_csv(csv_file_name):
     lastname_column = None
     for c in df.columns:
         if c.casefold() in (x.casefold() for x in possible_surname_fields):
-            print(f'"{c}" column present')
+            # print(f'"{c}" column present')
             firstname_column = c
             break
     for c in df.columns:
         if c.casefold() in (x.casefold() for x in possible_given_name_fields):
-            print(f'"{c}" column present')
+            # print(f'"{c}" column present')
             lastname_column = c
             break
 
@@ -94,81 +101,14 @@ def clean_non_canvas_csv(csv_file_name):
         raise ValueError("Cannot find appropriate column titles for names")
 
     # strip excess whitespace
-    df[firstname_column] = df[firstname_column].apply(lambda X: X.strip())
-    df[lastname_column] = df[lastname_column].apply(lambda X: X.strip())
+    df[firstname_column] = df[firstname_column].apply(lambda X: str(X).strip())
+    df[lastname_column] = df[lastname_column].apply(lambda X: str(X).strip())
 
     # concat columns to our preferred column
     df["studentName"] = df[firstname_column] + ", " + df[lastname_column]
 
     # just return the two relevant columns
     return df[["id", "studentName"]]
-
-
-def check_is_non_canvas_csv(csv_file_name):
-    """Read the csv file and check if id and appropriate student name exist.
-
-    1. Check if id is present.
-    2. Check if studentName is preset.
-    3. If not, check for given name and surname in the document.
-
-    Arguments:
-        csv_file_name (pathlib.Path/str): the csv file.
-
-    Returns:
-        bool
-    """
-
-    student_info_df = pandas.read_csv(csv_file_name, dtype="object")
-    print('Loading from non-Canvas csv file to check file: "{0}"'.format(csv_file_name))
-
-    # strip excess whitespace from column names
-    student_info_df.rename(columns=lambda x: x.strip(), inplace=True)
-
-    if "id" not in student_info_df.columns:
-        print('Cannot find "id" column')
-        print("Columns present = {}".format(student_info_df.columns))
-        return False
-
-    # if we have don't have  then we are good to go.
-    if "studentName" not in student_info_df.columns:
-
-        # we need one of some approx of last-name field
-        firstname_column_title = None
-        for column_title in student_info_df.columns:
-            if column_title.casefold() in (
-                x.casefold() for x in possible_surname_fields
-            ):
-                print('"{}" column present'.format(column_title))
-                firstname_column_title = column_title
-                break
-        if firstname_column_title is None:
-            print(
-                'Cannot find column to use for "surname", tried: {}'.format(
-                    ", ".join(possible_surname_fields)
-                )
-            )
-            print("Columns present = {}".format(student_info_df.columns))
-            return False
-
-        # we need one of some approx of given-name field
-        lastname_column_title = None
-        for column_title in student_info_df.columns:
-            if column_title.casefold() in (
-                x.casefold() for x in possible_given_name_fields
-            ):
-                print('"{}" column present'.format(column_title))
-                lastname_column_title = column_title
-                break
-        if lastname_column_title is None:
-            print(
-                'Cannot find column to use for "given name", tried: {}'.format(
-                    ", ".join(possible_given_name_fields)
-                )
-            )
-            print("Columns present = {}".format(student_info_df.columns))
-            return False
-
-    return True
 
 
 def clean_canvas_csv(csv_file_name):
@@ -187,55 +127,6 @@ def clean_canvas_csv(csv_file_name):
     student_info_df = student_info_df[["Student Number", "Student"]]
     student_info_df.columns = ["id", "studentName"]
     return student_info_df
-
-
-def check_is_canvas_csv(csv_file_name):
-    """Detect if a csv file is likely a Canvas-exported classlist.
-
-    Arguments:
-        csv_file_name (pathlib.Path/str): csv file to be checked.
-
-    Returns:
-        bool: True if we think the input was from Canvas, based on
-            presence of certain header names.  Otherwise False.
-    """
-    with open(csv_file_name) as csvfile:
-        csv_reader = csv.DictReader(csvfile, skipinitialspace=True)
-        csv_fields = csv_reader.fieldnames
-    return all(x in csv_fields for x in canvas_columns_format)
-
-
-def check_latin_names(student_info_df):
-    """Check if a dataframe has "studentName"s that encode to Latin-1.
-
-    Prints out a warning message for any that are not encodable.
-
-    Arguments:
-        student_info_df (pandas.DataFrame): with at least columns `id`
-            and `studentName`.
-
-    Returns:
-        bool: False if one or more names contain non-Latin characters.
-    """
-    # TODO - make this less eurocentric in the future.
-    encoding_problems = []
-    for index, row in student_info_df.iterrows():
-        try:
-            tmp = row["studentName"].encode("Latin-1")
-        except UnicodeEncodeError:
-            encoding_problems.append(
-                'row {}, number {}, name: "{}"'.format(
-                    index, row["id"], row["studentName"]
-                )
-            )
-
-    if len(encoding_problems) > 0:
-        print("WARNING: The following ID/name pairs contain non-Latin characters:")
-        for problem in encoding_problems:
-            print(problem)
-        return False
-    else:
-        return True
 
 
 def process_classlist_backend(student_csv_file_name):
@@ -262,12 +153,15 @@ def process_classlist_backend(student_csv_file_name):
     # we need to check it has the minimum information ie student name/id.
     # If not we will fail the process.
 
-    # First we check if this csv file is a Canvas output
-    if check_is_canvas_csv(student_csv_file_name):
+    # First we check if this csv file is a Canvas output - using the validator
+
+    vlad = PlomClasslistValidator()
+
+    if vlad.check_is_canvas_csv(student_csv_file_name):
         print("This file looks like it was exported from Canvas")
         student_info_df = clean_canvas_csv(student_csv_file_name)
         print("We have successfully extracted columns from Canvas data and renaming")
-    elif check_is_non_canvas_csv(student_csv_file_name):
+    elif vlad.check_is_non_canvas_csv(student_csv_file_name):
         print(
             "This file looks like it was not exported from Canvas; checking for the required information..."
         )
@@ -278,22 +172,28 @@ def process_classlist_backend(student_csv_file_name):
     else:
         raise ValueError("Problems with the supplied classlist. See output above.")
 
-    # Check characters in names are latin-1 compatible
-    if not check_latin_names(student_info_df):
-        print(">>> WARNING <<<")
-        print(
-            "Potential classlist problems",
-            "The classlist you supplied contains non-Latin characters - see console output above. "
-            "You can proceed, but it might cause problems later. "
-            "Apologies for the eurocentricity.",
-        )
-
-    # print("Saving to {}".format(outputfile))
-    # df.to_csv(outputfile, index=False)
     return student_info_df
 
 
-def get_demo_classlist():
+def print_classlist_warnings_errors(warn_err):
+    # separate into warn and err
+    warn = [X for X in warn_err if X["warn_or_err"] == "warning"]
+    err = [X for X in warn_err if X["warn_or_err"] != "warning"]
+    # sort by line number
+    warn.sort(key=lambda X: X["werr_line"])
+    err.sort(key=lambda X: X["werr_line"])
+
+    if warn:
+        print("Classlist validation warnings:")
+        for X in warn:
+            print(f"\tline {X['werr_line']}: {X['werr_text']}")
+    if err:
+        print("Classlist validation errors:")
+        for X in err:
+            print(f"\tline {X['werr_line']}: {X['werr_text']}")
+
+
+def get_demo_classlist(spec):
     """Get the demo classlist."""
     # Direct approach: but maybe I like exercising code-paths with below...
     # with resources.open_binary(plom, "demoClassList.csv") as f:
@@ -305,12 +205,19 @@ def get_demo_classlist():
     f = Path(tempfile.NamedTemporaryFile(delete=False, suffix=".csv").name)
     with open(f, "wb") as fh:
         fh.write(b)
-    C = process_classlist_file(f)
+
+    success, clist = process_classlist_file(f, spec, ignore_warnings=True)
+
+    if success is False:
+        raise ValueError(
+            f"Something has gone seriously wrong with the demo classlist - {clist}."
+        )
+
     f.unlink()
-    return C
+    return clist
 
 
-def process_classlist_file(student_csv_file_name):
+def process_classlist_file(student_csv_file_name, spec, *, ignore_warnings=False):
     """Get student names/IDs from a csv file.
 
     Student numbers come from an `id` column.  There is some
@@ -326,17 +233,40 @@ def process_classlist_file(student_csv_file_name):
 
     Arguments:
         student_csv_file_name (pathlib.Path/str): class info csv file.
+        spec (dict): validated test spec.
 
     Keyword Arguments:
-        demo (bool): if `True`, the filename is ignored and we use demo
-            data (default: `False`).
+        ignore_warnings (bool): if true, proceed with classlist
+            processing even if there are warnings.  Default False.
 
-    Return:
-        list: A list of dicts, each with `"id"` and `"studentName"`.
+    Return: tuple: if successful then "(True, clist)" where clist is a
+        list of dicts each with "id" and "studentName". On failure
+        "(False, warn_err)" where "warn_err" is a list of dicts of
+        warnings and errors. Each dict contains "warn_or_err" which is
+        'warning' or 'error', "werr_line" being the line number at
+        which the error occurs, and 'werr_text' being a string
+        describing the warning/error.
     """
     student_csv_file_name = Path(student_csv_file_name)
     if not student_csv_file_name.exists():
         raise FileNotFoundError(f'Cannot find file "{student_csv_file_name}"')
+
+    vlad = PlomClasslistValidator()
+    success, warn_err = vlad.validate_csv(student_csv_file_name, spec=spec)
+
+    if success is False:
+        # validation failed, return warning, error list
+        PlomClasslistValidator.print_classlist_warnings_errors(warn_err)
+        return (False, warn_err)
+
+    # validation passed but there are warnings
+    if warn_err:
+        print_classlist_warnings_errors(warn_err)
+        if ignore_warnings:
+            print("Continuing despite warnings")
+        else:
+            return (False, warn_err)
+
     df = process_classlist_backend(student_csv_file_name)
     # "records" makes it output a list-of-dicts, one per row
-    return df.to_dict("records")
+    return (True, df.to_dict("records"))
